lib: Fixing various warnings from the Visual Studio compiler about implicit casts between integer types.
The change for ThrowJVMError is probably an actual bug fix,
rather than just stopping the compiler complaining.

+++ src/m_jvmdebug.cc
@@ -62,7 +62,7 @@ void JVMModulesList::Parse(const Data &entry_packet)
 		Protocol::JVMModulesEntry *e = (Protocol::JVMModulesEntry *) ptr;
 
 		len = SB_JVMMODULES_ENTRY_HEADER_SIZE + be_btohs(e->sizename);
-		if( (count + len) > size )
+		if( (size_t)(count + len) > size )
 			break;
 
 		JVMModulesEntry entry;
@@ -128,7 +128,7 @@ void JVMThreadsList::Parse(const Data &entry_packet)
 		uint32_t *e = (uint32_t *) ptr;
 
 		len = sizeof(uint32_t);
-		if( (count + len) > size )
+		if( (size_t)(count + len) > size )
 			break;
 
 		JVMThreadsEntry entry;
@@ -253,7 +253,7 @@ void JVMDebug::Detach()
 }
 
 
-void JVMDebug::ThrowJVMError(const std::string &msg, uint8_t cmd)
+void JVMDebug::ThrowJVMError(const std::string &msg, uint16_t cmd)
 {
 	std::ostringstream oss;
 	oss << msg << ": unexpected packet command code: 0x"
